Changed arm motor to servo
