+++ src/main/java/frc/robot/commands/WristCommand.java
@@ -4,10 +4,9 @@
 
 package frc.robot.commands;
 import frc.robot.subsystems.WristSubsystem;
-import frc.robot.subsystems.WristSubsystem.IntakeState;
+import frc.robot.Constants.IntakeConstants.IntakeState;
 
 import edu.wpi.first.wpilibj2.command.Command;
-
 
 /* You should consider using the more terse Command factories API instead https://docs.wpilib.org/en/stable/docs/software/commandbased/organizing-command-based.html#defining-commands */
 public class WristCommand extends Command {
