+++ src/main/java/frc/robot/subsystems/WristSubsystem.java
@@ -7,55 +7,21 @@ package frc.robot.subsystems;
 import frc.robot.Constants.MotorConstants;
 
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import edu.wpi.first.math.controller.PIDController;
-
 import edu.wpi.first.wpilibj.Servo;
 
-import com.revrobotics.spark.SparkMax;
-import com.revrobotics.spark.SparkLowLevel.MotorType;
-import com.revrobotics.RelativeEncoder;
-
 public class WristSubsystem extends SubsystemBase {
+  private final Servo armServoMotor = new Servo(MotorConstants.kArmServoChannel);
+  private final Servo wristServoMotor = new Servo(MotorConstants.kWristServoChannel);
 
-  public enum IntakeState {
-    GroundIntake(0, 0), 
-    SourceIntake(45, 0), 
-    CoralScore(45, 90), 
-    Stow(90, 90);
-
-    public final double armAngle;
-    public final double wristAngle;
-
-    private IntakeState(double armAngle, double wristAngle) {
-      this.armAngle = armAngle;
-      this.wristAngle = wristAngle;
-    }
-  } 
-
-  private final SparkMax armMotor = new SparkMax(MotorConstants.kArmMotorId, MotorType.kBrushless);
-  private final RelativeEncoder armEncoder = armMotor.getEncoder();
-  
-  private final Servo servoWristMotor = new Servo(MotorConstants.kWristServoChannel);
-
-  private final PIDController armPID = new PIDController(.1, .1,0);
-  
   /** Creates a new WristSubsystem. */  
   public WristSubsystem() {}
 
-
-  public void setArmMotorSpeed(double speed) {
-    armMotor.set(speed);
-  }
-
-
   public void setArmMotorPosition(double desiredPosition) {
-    double currentPosition = armEncoder.getPosition();
-    double speed = armPID.calculate(currentPosition, desiredPosition);
-    setArmMotorSpeed(speed);
+    armServoMotor.setAngle(desiredPosition);
   }
 
   public void setWristMotorPostition(double desiredPosition) {
-    servoWristMotor.setAngle(desiredPosition);
+    wristServoMotor.setAngle(desiredPosition);
   }
 
   @Override
